requirements.txt
streamlit
python-dotenv
pymupdf
openai>=1.0.0
chromadb
tiktoken